feat: Process Markdown includes before building local search index
Closes #2812
